requirements.txt
cachetools==5.3.1
requests==2.31.0
python-dotenv==1.0.0
aiohttp==3.9.5
ftfy>=6.1.1
charset-normalizer>=3.0.0